Fix saving tenants in common DB2 owngroup.grpseqno value to match legacy application value.
